Refactor backend-server into modular packages
