+++ apps/backend/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "main": "dist/index.js",
   "scripts": {
-    "dev": "ts-node server/index.ts",
+    "dev": "ts-node -r tsconfig-paths/register server/index.ts",
     "build": "tsc",
     "lint": "eslint .",
     "check-types": "tsc --noEmit"
@@ -19,6 +19,7 @@
     "@types/node": "^22.15.18",
     "eslint": "^8.0.0",
     "ts-node": "^10.9.2",
+    "tsconfig-paths": "^4.2.0",
     "typescript": "^5.8.3"
   },
   "engines": {
