+++ apps/backend/tsconfig.json
@@ -1,9 +1,20 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../../tsconfig.json", 
   "compilerOptions": {
     "outDir": "dist",
-    "rootDir": "."
+    "rootDir": "./", 
+    "composite": true 
   },
-  "include": ["server", "prisma"],
-  "exclude": ["dist"]
+  "include": [
+    "server","**/*.ts" 
+   ],
+  "exclude": [
+    "dist",
+    "node_modules"
+  ],
+  "references": [
+    { "path": "../../packages/db/package.json" },
+    { "path": "../../packages/schema/package.json" },
+    { "path": "../../packages/services/package.json" } 
+  ]
 }
