+++ packages/db/prisma/proxy/find-many.proxy.ts
@@ -1,4 +1,4 @@
-import { Paginated } from "../../server/utils/paginated"; 
+import { Paginated } from "../../../../apps/backend/server/utils/paginated"; 
 import { PaginationData, ProxyFunctions } from "./types";
 
 const dateFields = ["createdAt", "updatedAt", "fecha"];
