+++ packages/db/repository/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@practicas/repository",
+  "version": "1.0.0",
+  "main": "index.ts"
+}
