+++ packages/db/repository/user.repository.ts
@@ -1,7 +1,7 @@
 import { prisma } from "./index";
-import { ProxyPrismaModel } from "../../prisma/proxy/prisma-proxy"; 
+import { ProxyPrismaModel } from "../prisma/proxy/prisma-proxy"; 
 import { Prisma, Usuario } from "@prisma/client";
-import { PaginationData } from "../../prisma/proxy/types";
+import { PaginationData } from "../prisma/proxy/types";
 
 export const createUserRepo = async (
   data: Prisma.UsuarioCreateInput,
