+++ pnpm-workspace.yaml
@@ -1,6 +1,7 @@
 packages:
   - "apps/*"
   - "apps/*/*"
+  - "packages/*"
 catalogs:
   prisma:
     prisma: latest
