+++ tsconfig.json
@@ -1,13 +1,29 @@
 {
-    "compilerOptions": {
-      "target": "ES2020",
-      "module": "CommonJS",
-      "rootDir": "./",              
-      "outDir": "./dist",           
-      "esModuleInterop": true,
-      "strict": true,
-      "moduleResolution": "node",
-      "skipLibCheck": true
-    },
-    "include": ["server", "apps/client"]
-  }
+  "compilerOptions": {
+    "target": "ES2020",
+    "module": "CommonJS",
+    "rootDir": "./",
+    "outDir": "./dist",
+    "esModuleInterop": true,
+    "strict": true,
+    "moduleResolution": "node",
+    "forceConsistentCasingInFileNames": true,
+    "skipLibCheck": true,
+    "baseUrl": ".",
+    "paths": {
+      "@db/*": [
+        "packages/db/*"
+      ],
+      "@schema/*": [
+        "packages/schema/*"
+      ],
+      "@services/*": [
+        "packages/services/*"
+      ]
+    }
+  },
+  "include": [
+    "server",
+    "apps/client"
+  ]
+}
